Move Questions component from routes dir to components/
