+++ src/main/webapp/src/components/Questions/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "Questions",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./Questions.js"
+}
