+++ src/main/webapp/src/components/Questions/question-icon.png
